Fix: Add int

+++ app/Http/Controllers/STXI/EMS2/deliveryMethodToPSIController.php
@@ -485,7 +485,7 @@ class deliveryMethodToPSIController extends BaseController
     {
         ini_set('max_execution_time', '300');
 
-        $date_to = date('d', strtotime($date)) == 1 ? date('Y-m-d', strtotime($date . "-1 days")) : date('Y-m-d');
+        $date_to = (int)date('d', strtotime($date)) == 1 ? date('Y-m-d', strtotime($date . "-1 days")) : date('Y-m-d');
         if (!empty($item)) {
             $query = "SET NOCOUNT ON;EXEC Z_STXI_GET_CPO_DLV_STXI_ITEC @date_start = '" . date('Y-m-01', strtotime($date)) . "', @date_to = '" . $date_to . "', @model = '" . $item . "'";
         } else {
